Cleanup
Move include files out of the iCub folder
