+++ src/yarpdataplayer-qt/mainwindow.h
@@ -13,7 +13,7 @@
 #include <string>
 #include <map>
 #include <fstream>
-#include "iCub/utils.h"
+#include "utils.h"
 #include <yarp/os/ResourceFinder.h>
 #include <yarp/os/Value.h>
 #include <yarp/os/Bottle.h>
