+++ src/yarpdataplayer-qt/res.qrc
@@ -1,17 +1,17 @@
-<RCC>
-    <qresource prefix="/">
-        <file>images/forward.png</file>
-        <file>images/pause.png</file>
-        <file>images/play.png</file>
-        <file>images/rewind.png</file>
-        <file>images/stop.png</file>
-        <file>images/splash/bg1.png</file>
-        <file>images/splash/bg2.png</file>
-        <file>images/splash/bg3.png</file>
-        <file>images/splash/bg4.png</file>
-        <file>images/splash/bg5.png</file>
-        <file>images/splash/bg6.png</file>
-        <file>images/splash/bg7.png</file>
-        <file>images/splash/bg8.png</file>
-    </qresource>
+<!DOCTYPE RCC><RCC version="1.0">
+<qresource>
+    <file>images/forward.png</file>
+    <file>images/pause.png</file>
+    <file>images/play.png</file>
+    <file>images/rewind.png</file>
+    <file>images/stop.png</file>
+    <file>images/splash/bg1.png</file>
+    <file>images/splash/bg2.png</file>
+    <file>images/splash/bg3.png</file>
+    <file>images/splash/bg4.png</file>
+    <file>images/splash/bg5.png</file>
+    <file>images/splash/bg6.png</file>
+    <file>images/splash/bg7.png</file>
+    <file>images/splash/bg8.png</file>
+</qresource>
 </RCC>
